+++ app/src/main/java/com/kuhrusty/morbadscorepad/DangerDeckActivity.java
@@ -39,6 +39,7 @@ public class DangerDeckActivity extends AppCompatActivity {
     private MediaPlayer mediaPlayer;
 
     private DeckState<Danger> deck;
+    private int keepShuffleCount = 5;  //  just a guess at what would be good
 
     private boolean needConfirmDangerUpdate = false;
 
@@ -162,7 +163,10 @@ Log.w(LOGBIT, "need to load deck state");
     public void doDraw(View view) {
         killSound();
         Danger card = deck.draw();
-        if (card == null) deck.shuffle();
+        if (card == null) {
+            deck.shuffle();
+            deck.trimLog(keepShuffleCount);
+        }
         needConfirmDangerUpdate = confirmPref && (card != null);
         updateUI(true);
         if ((card == null) || (audioPref == null)) return;  //  no audio
@@ -186,7 +190,10 @@ resID = "danger_" + voicePref + "_" + card.getID();
      */
     public void doSecondaryDraw(View view) {
         killSound();
-        if (deck.draw() == null) deck.shuffle();
+        if (deck.draw() == null) {
+            deck.shuffle();
+            deck.trimLog(keepShuffleCount);
+        }
         updateUI(true);
     }
 
@@ -206,6 +213,7 @@ resID = "danger_" + voicePref + "_" + card.getID();
 
     public void doShuffleDrawPile(View view) {
         deck.shuffleDrawPile();
+        deck.trimLog(keepShuffleCount);
         updateUI(false);
     }